new_requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
boto3==1.17.110
botocore==1.20.110
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
enum-compat==0.0.3
gast==0.2.2
gitdb==4.0.7
GitPython==3.1.18
google-pasta==0.2.0
grpcio==1.38.1
h5py==2.10.0
idna==2.10
importlib-metadata==4.6.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkg-resources==0.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==4.0.1
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
s3transfer==0.4.2
Send2Trash==1.7.1
six==1.16.0
smmap==4.0.0
streamlit==0.70.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.5.0
